Improve network connection system
UDP is somewhat unreliable and even less reliable when using
multicast packets. This patch changes the networking model from a
completely peer-to-peer setup where the role of a "server" switches
to whoever last initiated a scan and instead has clients send a
join message over multicast. All other messages are sent directly
between the clients with one device taking on the role of serving
the lobby and sending out the player list. Clients can attempt to
join multiple times and even connect directly to the lobby by
entering an IP if all else fails.

We are still using a peer-to-peer setup for all other communication
however. Clients will still send messages directly to each other
with no device being the center or server.

Added a check to ensure that Wi-Fi is connected.

Fixed a deadlock that sometimes occurred if a user tapped the
join / leave button twice quickly and the sockets did not stop
listening before tapping join again.

I also started working a feature to enter player names, but this
feature is not fully implemented yet.