Make up to 63 players possible with a simple variable change
Made the max player value a static final so it can be easily
changed and tweaked hit detection code to handle the conversion of
signed bytes to unsigned ints so that we can properly send hit data
to networked players. I am still limiting the GUI selection to 16
players / teams because I don't think most people need to be able
to select more than 16 and the current input mechanism would be
cumbersome to use with 63 options.