Set game limits
Players may now set one or more limits to the game.
-Time limit
-Lives / respawn limit
-Score limit (network games only)

In networked games, the game limits will be set by the lobby creator
and shared to clients when the client joins the lobby. A network
game will automatically end when one player reaches the score limit.
Players that reach the elimination limit will be out of the game,
but the other players will be able to continue playing. The last
player standing will have to manually end the game.