Add player names support in network games
The GUI will now request a player name once when enabling network
features. Names will be displayed for people you are hitting, people
that are hitting you, people that eliminated you, and when you score
a point.

A team score is now displayed. When someone on your team scores a
point, the team score will be updated. If a score limit is set,
the game will end when one team reaches the score limit.

Added a sound that plays when you score a point in network games.

Fixed a few bugs (and probably created a couple of new bugs).