UI rework, add indoor/outdoor modes, detect rifle/pistol
Get rid of separate network buttons and use a popup menu to clean
up the UI a bit.

Found configuration for and implemented a button to select indoor
and outdoor modes. Outdoor mode can be selected with and without
the close range broad area IR cone support. This item is a saved
preference that will reload its previous state each time the app
is opened.

Implemented a method that should hopefully detect if the user has
connected a rifle or a pistol and adjust the battery gauge for
the appropriate blaster.

Implemented a characteristic read queue because it was needed to
allow reading the blaster's ID characteristic during blaster
connection so we can determine if we have a rifle or a pistol.