Add option to reconnect only to previously paired weapon
Save the MAC address of the last blaster paired. If a blaster has
previously been paired, present an option to reconnect to the same
weapon only. This may be used to reconnect to only that same weapon
to speed up repairing in large groups and prevent phones from
pairing with the wrong blaster.