Add QR code scanning to connect blasters
I generated the QR code at https://www.the-qrcode-generator.com/
and saved it to SVG. I opened it in GIMP using 100 pixels per
inch and specified a 1 inch sized image. I printed the result,
cut it out, and used some clear packing tape to get the QR code
onto the side of the blaster.